Add Home component. Update styles.
