+++ styles/_index.scss
@@ -10,7 +10,6 @@
 }
 
 .cover {
-  // background-image: linear-gradient(45deg,#00aeef 0%,#096fb9 100%);
   overflow: hidden;
   position: relative;
 }
@@ -37,40 +36,21 @@
   display: flex;
   justify-content: center;
   align-items: center; 
-  // right: 5em;
-  // top: -5em;
+  
+  .image {
+    position: absolute;
+    height: 700px;
 
-  // .hero-section {
-  //   h2 {
-  //     color: white;
-  //     font-weight: bold;
-  //     margin-bottom: 10px;
-  //   }
-  
-  //   perspective: 1000px;
-  //   color: white;
-  //   right: 1em;
-  
-    .image {
-      // width: auto;
-      // max-width: 120%;
-      position: absolute;
-      height: 700px;
-  
-      &.fg {
-        animation-duration: 15s;
-        animation-iteration-count: infinite;
-      }
-  
-      &.bg {
-        opacity: 15%;
-      }
-    // }  
+    &.fg {
+      animation-duration: 15s;
+      animation-iteration-count: infinite;
+    }
+
+    &.bg {
+      opacity: 15%;
+    } 
   }
 }
-  
-
-  
 
 .hero-welcome-headline {
   margin-bottom: 20px;
@@ -81,7 +61,6 @@
     font-size: 60px;
     margin-top: 80px;
     font-weight: bold;
-    // margin-bottom: 20px
   }
 }
 
@@ -91,31 +70,14 @@
 
   h3 {
     color: white;
-    // text-transform: uppercase;
     font-size: 24px;
     margin-top: 20px;
   }
 }
-  
-// .hero-welcome-bio {
-//   margin-top: 20px;
-//   color: white;
-
-//   .fa-stack {
-//     font-size: 28px;
-//   }
-
-//   .list-inline-item {
-//     > a {
-//       color: #373737;
-//     }
-//   }
-// }
 
 .hero-welcome-wrapper {
   justify-content: flex-end;
 }
-
 
 @media (max-width: 991px) {
   .hero-section {
@@ -130,30 +92,3 @@
     margin-top: 0px;
   }
 }
-
-// .background-image {
-//   position: absolute;
-//   width: 1762px;
-//   height: 493px;
-// }
-
-// .shadow-custom {
-//   width: 100%;
-//   height: 100%;
-//   position: absolute;
-//   top: 0;
-//   box-shadow: -17px 13px 41px rgba(13,78,158,0.6);
-//   border: 2px solid #3bace5;
-
-//   .shadow-inner {
-//     width: 100%;
-//     height: 100%;
-//     box-shadow: -17px 13px 13px rgba(0,0,0,0.09);
-//   }
-// }
-
-// .shadow-custom-2 {
-//   border: 2px solid #f37d49;
-//   box-shadow: -17px 13px 41px rgba(13,78,158,0.3);
-// }
-
