+++ styles/components/about/_aboutme.scss
@@ -10,8 +10,12 @@
   white-space: pre-wrap;
 
   .header {
-      font-size: 28px;
+      font-size: 36px;
       font-weight: bold;
       margin-bottom: 25px;
+  }
+
+  .description {
+    font-size: 22px;
   }
 }
