+++ styles/components/about/_skillcard.scss
@@ -1,8 +1,6 @@
 .skill-card-container {
   height: 9em;
   width: 10em;
-  // border: 1px solid #CCC;
-  // margin: 40px 0;
   perspective: 600px;
 
   transition: transform 1s;
@@ -19,25 +17,14 @@
   position: relative;
   border: 2.7px solid $cornflower-faded;
   border-radius: 10px;
-  
 
   &:hover {
-    //transform: scale(1.05);
-    //border-radius: 10px;
     @include box-shadow(0px 0px 20px 0px $dodger-blue-faded);
-    //cursor: pointer;
   }
 
   &.is-flipped {
     transform: rotateY(180deg);
-
-    // &:hover {
-    //   transform: scale(1.05);
-    // }
   }
-
-
-
 }
 
 .hovered {
@@ -45,7 +32,6 @@
   transform-style: preserve-3d;
   transform: scale(1.05);
   border-radius: 10px;
-  // @include box-shadow(0px 0px 20px 0px $dodger-blue-faded);
   cursor: pointer;
 }
 
@@ -53,36 +39,26 @@
   position: absolute;
   width: 100%;
   height: 100%;
-  // line-height: 260px;
   color: white;
   text-align: center;
-  // font-weight: bold;
-  // font-size: 40px;
   -webkit-backface-visibility: hidden;
   backface-visibility: hidden;
 
   display: flex;
   justify-content: center;
   align-items: center;
- // height: 100%;
-
  
-
-
   &.front {
-    // background: red;
     font-size: 75px;
   }
 
   &.back {
-    background: $matisse-faded;
+    background: #2f82e266; //#0c58af8f;
     transform: rotateY(180deg);
     border-radius:  8px;
   }
 
   .card-details {
-    // display: flex;
-
     font-size: 24px;
     
     .years {
