+++ styles/components/home/_home.scss
@@ -1,0 +1,3 @@
+.home-page {
+ padding: 0px !important;
+}
