+++ styles/main.scss
@@ -8,6 +8,8 @@
 @import './pages/about';
 @import './pages/home';
 
+@import './components/home/home';
+
 @import './components/about/skillcard';
 @import './components/about/aboutme';
