+++ styles/pages/_about.scss
@@ -1,4 +1,3 @@
-// MIXINS
 @mixin background-cover() {
   -webkit-background-size: cover;
   -moz-background-size: cover;
@@ -15,7 +14,7 @@
 }
 
 @mixin animation-fill-mode($animationMode) {
-  -webkit-animation-fill-mode: $animationMode;  /* this makes sure that after animation is done we remain at the last keyframe value (opacity: 1)*/
+  -webkit-animation-fill-mode: $animationMode;
   -moz-animation-fill-mode: $animationMode;
   animation-fill-mode: $animationMode;
 }
