+++ styles/pages/_home.scss
@@ -1,8 +1,7 @@
 .technical-experience {
   color: $white;
-  margin-top: 2em; // 100px;
+  margin-top: 2em;
   text-align: center;
-  // margin-top: 380px;
 
   .tagline {
     margin: 0 auto 25px auto;
@@ -39,7 +38,7 @@
 }
 
 .hero-resume {
-  margin-top: 4em; // 1em;
+  margin-top: 4em;
   display: flex;
   
   .btn {
