+++ styles/pages/_resume.scss
@@ -55,10 +55,10 @@
     
       .card-text {
         line-height: 18px;
-        font-size: 12px;
+        font-size: 18px;
         display: inline-block;
         font-weight: 600;
-        margin-bottom: 3px;
+        margin-bottom: 10px;
         display: flex;
       
         .location {
@@ -69,7 +69,7 @@
       
         .company {
           float: left;
-          font-size: 14px;
+          font-size: 20px;
         }
       
         .title {
@@ -84,7 +84,7 @@
       }
     
       ul {
-        font-size: 12px;
+        font-size: 18px;
         padding-left: 16px;
         text-align: justify;
         margin-top: 15px;
